admin users

+++ controllers/contacts/index.js
@@ -4,4 +4,5 @@ const { add } = require("./add");
 const { deleteById } = require("./deleteById");
 const { putById } = require("./putById");
 const { patchFavorite } = require("./patchFavorite");
-module.exports = { getAll, getById, add, deleteById, putById, patchFavorite };
+const { getAdminUsers } = require("./getAdminUsers")
+module.exports = { getAll, getById, add, deleteById, putById, patchFavorite, getAdminUsers };
